requirements.txt
opencv-python
numpy
tensorflow
scikit-learn
flask
pillow
mtcnn
faiss-cpu
PyYAML
